+++ src/movie.h
@@ -210,7 +210,7 @@ public:
 	//whether microphone is enabled
 	bool microphone;
 
-	int getNumRecords() { return records.size(); }
+	int getNumRecords() { return (int)records.size(); }
 
 	int RAMInitOption, RAMInitSeed;
